+++ src/Bullet/Bullet.h
@@ -4,7 +4,7 @@ class Bullet
 {
 public:
 	Bullet();
-	Bullet(float x, float y, float dir[2], double bulletSpeed, float bulletDamage);
+	Bullet(float x, float y, float dir[2], float bulletSpeed, float bulletDamage);
 	void update(); //this function will draw and move bullet
 	void draw();
 	void move();
